![image](https://github.com/user-attachments/assets/decdc9ef-0763-4d1c-b3ae-e452b5155c53)
(He's a bit dumb because he was made with `--fast` and lite patch.)

This pull:
* Enables support with third-party checkpoints. Tested with https://civitai.com/models/686814/jib-mix-flux and https://civitai.com/models/673188/acorn-is-spinning-flux
* Adds a `FluxModSamplerWrapper` node that enables using FluxMod with samplers like `SamplerCustom`.
* Refactors the model loading to use the existing Flux modules when possible and passes `operations` through. Makes FluxMod work with ComfyUI's `--fast` option (though it seems to reduce quality a fair amount). Also appears to fix a problem where some tensors would be on the wrong device when there was a lot of memory pressure (I could reliably reproduce it doing a 3x upscale from 1024x1024).

Tested to work on my system (Linux, 4060Ti) with the finetunes listed above, normal Dev, Schnell, `--fast` or not and lite patch enabled or not.

The first two bullet points in the list of changes are pretty minimal from a code standpoint. This pull could be split up to separate out the loader/layer refactoring since it is fairly complex.