@blepping 2025-01-05
I did a bit of performance testing to make sure these changes didn't have a negative effect:

Tested|Speed
-|-
No FluxMod|2.45s/it
No FluxMod, `--fast`|1.72s/it
Original FluxMod|2.63s/it
Original FluxMod, `--fast`|No change
This pull|2.63s/it
This pull, `--fast`|1.71s/it

Tests done on the jibMix model, loaded as `float8_e4m3fn`, 1024x1024, 40 steps on a 4060Ti 16GB. Setting the activation type to `bf16` or `fp16` had no effect on performance in the original version or with my changes.